requirements.txt
fastapi
uvicorn[standard]
python-multipart
sentence-transformers
faiss-cpu
openai
numpy
PyPDF2
python-docx
python-pptx
openpyxl
striprtf
beautifulsoup4
